requirements.txt
export
git+https://github.com/dvska/gdata-python3#egg=gdata
python-Levenshtein
pymysql
netaddr
paramiko>=1.16.0
python-dateutil
PyYAML>=3.11
redis
# scipy - just use the package.
simplemediawiki
